chore: update config

+++ nvim/lua/plugins/toggleterm.lua
@@ -1,13 +1,16 @@
-vim.cmd [[
-		let &shell = executable('pwsh') ? 'pwsh' : 'powershell'
-		let &shellcmdflag = '-NoLogo -ExecutionPolicy RemoteSigned -Command [Console]::InputEncoding=[Console]::OutputEncoding=[System.Text.UTF8Encoding]::new();$PSDefaultParameterValues[''Out-File:Encoding'']=''utf8'';Remove-Alias -Force -ErrorAction SilentlyContinue tee;'
-		let &shellredir = '2>&1 | %%{ "$_" } | Out-File %s; exit $LastExitCode'
-		let &shellpipe  = '2>&1 | %%{ "$_" } | tee %s; exit $LastExitCode'
-		set shellquote= shellxquote=
-]]
+if vim.fn.has 'win32' then
+	vim.cmd [[
+			let &shell = executable('pwsh') ? 'pwsh' : 'powershell'
+			let &shellcmdflag = '-NoLogo -ExecutionPolicy RemoteSigned -Command [Console]::InputEncoding=[Console]::OutputEncoding=[System.Text.UTF8Encoding]::new();$PSDefaultParameterValues[''Out-File:Encoding'']=''utf8'';Remove-Alias -Force -ErrorAction SilentlyContinue tee;'
+			let &shellredir = '2>&1 | %%{ "$_" } | Out-File %s; exit $LastExitCode'
+			let &shellpipe  = '2>&1 | %%{ "$_" } | tee %s; exit $LastExitCode'
+			set shellquote= shellxquote=
+	]]
+end
+
 
 require('toggleterm').setup{
-	open_mapping = '<C-t>',
+	open_mapping = '<C-\\>',
 	star_in_insert = true,
 	direction = "float",
 }
